+++ modules/main/src/main/java/com/nisovin/shopkeepers/config/Settings.java
@@ -210,7 +210,8 @@ public class Settings extends Config {
 					"BREEZE", // MC 1.21
 					"CREAKING", // MC 1.21.4
 					"HAPPY_GHAST", // MC 1.21.6
-					"COPPER_GOLEM" // MC 1.21.9
+					"COPPER_GOLEM", // MC 1.21.9
+					"MANNEQUIN" // MC 1.21.9
 			), String::compareTo)
 	);