+++ modules/v1_21_R8/src/main/java/com/nisovin/shopkeepers/compat/v1_21_R8/CompatProviderImpl.java
@@ -25,17 +25,21 @@ import org.bukkit.entity.Cow;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Golem;
 import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Mannequin;
 import org.bukkit.entity.Mob;
 import org.bukkit.entity.Pig;
 import org.bukkit.entity.Player;
+import org.bukkit.entity.Pose;
 import org.bukkit.entity.Salmon;
 import org.bukkit.entity.Villager;
 import org.bukkit.event.block.BlockExplodeEvent;
 import org.bukkit.event.entity.EntityExplodeEvent;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.MainHand;
 import org.bukkit.inventory.Merchant;
 import org.bukkit.inventory.MerchantInventory;
+import org.bukkit.profile.PlayerProfile;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
@@ -503,5 +507,25 @@ public final class CompatProviderImpl implements CompatProvider {
 
 	public void setCopperGolemNextWeatheringTick(Golem golem, int tick) {
 		((CopperGolem) golem).setNextWeatheringTick(tick);
+	}
+
+	public void setMannequinHideDescription(LivingEntity mannequin, boolean hideDescription) {
+		((Mannequin) mannequin).setHideDescription(hideDescription);
+	}
+
+	public void setMannequinDescription(LivingEntity mannequin, @Nullable String description) {
+		((Mannequin) mannequin).setDescription(description);
+	}
+
+	public void setMannequinMainHand(LivingEntity mannequin, MainHand mainHand) {
+		((Mannequin) mannequin).setMainHand(mainHand);
+	}
+
+	public void setMannequinPose(LivingEntity mannequin, Pose pose) {
+		((Mannequin) mannequin).setPose(pose);
+	}
+
+	public void setMannequinProfile(LivingEntity mannequin, @Nullable PlayerProfile profile) {
+		((Mannequin) mannequin).setPlayerProfile(profile);
 	}
 }